Address review: inline test methods, drop mixin pattern
Per review feedback from jprakash-db:
- Remove mixin classes (LargeWideResultSetMixin, etc) — inline the
  test methods directly into the test classes in test_driver.py
- Remove backward-compat LargeQueriesMixin alias (nothing uses it)
- Rename _LargeQueryRowHelper — replaced entirely by inlining
- Convert large_queries_mixin.py to just a fetch_rows() helper function
